Updated Test for Travis

+++ Transformation/transformation/src/main/scala/export/TileVisualizer.scala
@@ -114,6 +114,9 @@ class TileVisualizer {
     }
 
     val fileName = (new PathFormatter).getDirectory(para, extra) +tile.rows+"______"+para.weightMatrix+"r_"+para.weightRadius+"_"+tile.cols+"Time_"+ LocalDateTime.now().formatted("HH_mm" ) + ".png"
+    if(!new File(fileName).exists()){
+      return 
+    }
     val fos = new FileOutputStream(fileName);
     ImageIO.write(bfI, "PNG", fos);
     fos.close();
